Feat/dead pixel (#34)
* feat(settings): split up settings page

* feat(dead-pixel): extend storage to support flag

* feat(dead-pixel): add barebone experience

* feat(dead-pixel): update readme
