+++ src/features/dead_pixel/index.ts
@@ -1,0 +1,1 @@
+export { default as DeadPixelHost } from './DeadPixelHost';
